chore: home moved to /home
