+++ src/app/[locale]/home/components/Buttons/ApiInfo.tsx
