+++ src/app/[locale]/home/components/Buttons/SessionInfo.tsx
