+++ src/app/[locale]/layout.tsx
@@ -1,6 +1,6 @@
 import Body from './components/Body'
-import SignButton from './components/Buttons/Sign'
 import Menu from './components/Menu'
+import SignButton from './components/Sign'
 import { Providers } from './providers'
 
 import {
@@ -38,7 +38,7 @@ export default async function RootLayout({
                 <Menu.Button
                   icon={<House weight="bold" />}
                   text={tSidebar('home')}
-                  path="/"
+                  path="/home"
                 />
                 <Menu.Button
                   icon={<Brain weight="bold" />}
