+++ src/app/page.tsx
@@ -3,7 +3,7 @@ import options from '@/common/auth'
 import { redirect } from 'next/navigation'
 import { getServerSession } from 'next-auth'
 
-export default async function Home() {
+export default async function Index() {
   const session = await getServerSession(options)
   session ? redirect('/home') : redirect('/signin')
 }
